Add files via upload
User functions to calculate the box dimension of the logistical map. The box counting dimension is a measure of how much of the available phase space the map will utilize for a given parameter r and initial x, for the logistical map. The available values the map can take are between 0 and 1., so the box dimension of the logistic map will demonstrate how much of the domain (values of x between 0 and 1) will be visited by the map given enough iterations. A value of 1 indicates that the whole domain can potential be visited at some point. A value of 0.5 means about half of the domain can be visited for those particular value of x0 and r. It must be noted that a value less than 1 doesn't imply chaos is not possible. A sub range such as 0-0.5 can still have infinite points within it and thus contain chaotic trajectories. Stable orbits or periodic orbits should have a dimension of 0 as they only utilize a small number of points.